Refacto: consts & copies & pointers & vector-bound
Preperation for the implementation of the new pause-moving logic.

For verifying the new pause location, we will re-launch the TSPTWSolver
because this model already has all the constraints, relations, types,
time-windows, etc. and checkassingment() function is very-fast (<millisecods).

To do this, TSPTWDataDT needed refactoring so that we will not end up
with dangling pointers or free objects allocated in a different context
in the wrong or duplicate unnecessary parts (like matrix).