Fix maximum day lapse relation
Due to day_to_vehicle_evaluator returning the index of the first
vehicle of the corresponding day, some feasible solutions were getting
rejected by the or-tools model. So the fix is to get the index of the
first vehicle of the next day and decrease it by one (to have the index
of the last vehicle of the corresponding day.